+++ R/tidy.R
@@ -1,10 +1,10 @@
 #' Turn sdmTMB model output into a tidy data frame
 #'
 #' @param x Output from [sdmTMB()].
-#' @param effects A character vector including one or more of "fixed"
-#'   (fixed-effect parameters), "ran_pars" (standard deviations, spatial range,
-#'   and other random effect and dispersion terms), or "ranef" (individual random
-#'   intercepts, if included -- behaves like `ranef()`).
+#' @param effects A character value. One of `"fixed"` ('fixed' or main-effect
+#'   parameters), `"ran_pars"` (standard deviations, spatial range, and other
+#'   random effect and dispersion-related terms), or `"ran_vals"` (individual
+#'   random intercepts, if included; behaves like `ranef()`).
 #' @param conf.int Include a confidence interval?
 #' @param conf.level Confidence level for CI.
 #' @param exponentiate Whether to exponentiate the fixed-effect coefficient
@@ -17,7 +17,7 @@
 #' Follows the conventions of the \pkg{broom} and \pkg{broom.mixed} packages.
 #'
 #' Currently, `effects = "ran_pars"` also includes dispersion-related terms
-#' (e.g., `phi`), which are not actually random effects.
+#' (e.g., `phi`), which are not actually associated with random effects.
 #'
 #' Standard errors for spatial variance terms fit in log space (e.g., variance
 #' terms, range, or parameters associated with the observation error) are
@@ -35,8 +35,15 @@
 #' tidy(fit)
 #' tidy(fit, conf.int = TRUE)
 #' tidy(fit, "ran_pars", conf.int = TRUE)
+#'
+#' pcod_2011$fyear <- as.factor(pcod_2011$year)
+#' fit <- sdmTMB(density ~ poly(depth_scaled, 2, raw = TRUE) + (1 | fyear),
+#'   data = pcod_2011, mesh = pcod_mesh_2011,
+#'   family = tweedie()
+#' )
+#' tidy(fit, "ran_vals")
 
-tidy.sdmTMB <- function(x, effects = c("fixed", "ran_pars", "ranef"), model = 1,
+tidy.sdmTMB <- function(x, effects = c("fixed", "ran_pars", "ran_vals"), model = 1,
                  conf.int = FALSE, conf.level = 0.95, exponentiate = FALSE, ...) {
   effects <- match.arg(effects)
   assert_that(is.logical(exponentiate))
@@ -248,7 +255,10 @@ tidy.sdmTMB <- function(x, effects = c("fixed", "ran_pars", "ranef"), model = 1,
 
   # random intercepts
   n_re_int <- length(x$split_formula[[model]]$reTrmFormulas)
-  if(n_re_int > 0) {
+  if (n_re_int == 0 && effects == "ran_vals") {
+    cli::cli_abort("effects = 'ran_vals' currently only works with random intercepts (e.g., `+ (1 | g)`).")
+  }
+  if (n_re_int > 0) {
     out_ranef <- list()
     re_est <- as.list(x$sd_report, "Estimate")$RE
     re_ses <- as.list(x$sd_report, "Std. Error")$RE
@@ -285,12 +295,18 @@ tidy.sdmTMB <- function(x, effects = c("fixed", "ran_pars", "ranef"), model = 1,
   out <- unique(out) # range can be duplicated
   out_re <- unique(out_re)
 
-  if (effects == "fixed") {
-    return(out)
-  } else if (effects == "ranef") {
-    return(out_ranef)
+  if (requireNamespace("tibble", quietly = TRUE)) {
+    frm <- tibble::as_tibble
   } else {
-    return(out_re)
+    frm <- as.data.frame
+  }
+
+  if (effects == "fixed") {
+    return(frm(out))
+  } else if (effects == "ran_vals") {
+    return(frm(out_ranef))
+  } else {
+    return(frm(out_re))
   }
 }